Fix linux build (#101)

+++ src/json/json.c
@@ -1343,7 +1343,7 @@ json_unescape (const char *text)
 					}
 					else
 					{
-						fprintf (stderr, "JSON: unsupported unicode value: 0x%lX\n", unicode);
+						fprintf (stderr, "JSON: unsupported unicode value: 0x%lX\n", (long unsigned int)unicode);
 					}
 				}
 				break;
